pyproject.toml
[project]
name = "gan-from-scratch-pytorch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "ipykernel>=6.30.1",
    "matplotlib>=3.9.4",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
]
